worker/requirements.txt
yfinance==0.2.38
pandas==2.2.1
psycopg2-binary==2.9.9